Cost関連の調整対応
* UnitDefinitionと、EnemyStatusでCost管理が二重になっていたので直した
* Costをfloatでなく、intで管理するようにした

+++ Assets/Script/Interface/IEconomy.cs
@@ -5,14 +5,14 @@ public interface IEconomy
     // インターフェースはアクセスレベルは基本的に使わない
     // (ごく例外もあるが、基本publicとして使う）
     // IEconomyを持つクラスは、これらの実装を必ず持つというルール
-    float CurrentMoney { get; }
+    int CurrentMoney { get; }
 
-    event Action<float> MoneyChanged;
+    event Action<int> MoneyChanged;
     event Action<string> OnInsufficientFunds; // 支払い失敗のイベント
 
     bool TrySpend(int cost);
 
-    void Refund(float cost);
+    void Refund(int cost);
 
-    void AddMoney(float value);
+    void AddMoney(int value);
 }
